API: Plot inverted geometry with inverted colors
It was hard to see cut-outs because for example, if you cut a
polygon out of a polygon, it looks like two stacked polygons
which are the same color. That made it hard to see that in fact,
there was a negative space. By switching the colors of inverse
geometry to inverse colors, the cut-outs are much more visible.