API: Change coverage glyphs to PieGlyph
The coverage Glyph was not expressive enough because it was an
ellipse which meant it could only express coverage from two
perpendicular directions. I also tried star glyphs but those are
harder to see and when the linewidth is larger than 0, there are
rendering artefacts when the Glyphs are small. These new PieGlyphs
are better because they more accruately express the bin directions
and they do not have rendering artefacts.