API: Flip pyplot instead of numpy for plotting
Earlier I had flipped the output of plot.discrete_phantom to match
the matplotlib plot of the geometry patches for plot.sidebyside.
matplotlib.patches uses the horizontal as x and vertical as y, but
when you use imshow, the vertical axis is the first axis. By switching
the discrete output, reconstructions were always the transpose of the
ground truth. This was annoying, so instead of changing the discrete
data, now the patches are inverted to match. This is less data
intensive because now we don't have to transpose a bunch of matricies
in order to make quality assessment.