API: Refactor SiemensStar to only count materials as blades
The n_sectors param was a little wierd because odd numbers of
blades would not be evenly spaced. This is because dark spaces
between blades were counted as blades. This fixes that and adds
a two functions to compute the spatial frequency at a given
radius or find a radius that computes a frequency.