REF: Automatic number of samples for get_line_at_radius
Because:
- The number of samples could be set too low

Sampling the line at a radius with less points than PI * diameter
of the circle is insufficient to capture high frequency information
that we are tying to characterize. The number of angles is now
automatically at least PI * diameter. Technically, the smaller radii
are over sampled, but that should be fine.